refactor(utils): move image operation functions to `imgutil` package
