+++ filters/horizontal_flip/serial.go
@@ -3,7 +3,7 @@ package horizontal_flip
 import (
 	"image"
 
-	"dougdomingos.com/image-filters/utils"
+	"dougdomingos.com/image-filters/filters/imgutil"
 )
 
 // serialHorizontalFlip applies the horizontal flip filter to the entire image
@@ -19,7 +19,7 @@ func serialHorizontalFlip(img *image.RGBA) {
 			leftOffset := rowStart + (bounds.Min.X+deltaX)*4
 			rightOffset := rowStart + (bounds.Max.X-1-deltaX)*4
 
-			utils.SwapPixels(img, leftOffset, rightOffset)
+			imgutil.SwapPixels(img, leftOffset, rightOffset)
 		}
 	}
 }
