+++ filters/imgutil/pixel_operations.go
@@ -1,4 +1,4 @@
-package utils
+package imgutil
 
 import "image"
